<FIx> hp 수정

+++ Assets/Scripts/Actor/Actor.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Actor : MonoBehaviour
 {
-    [Range(1, 100)]public int hp = 100;
+    [Range(1, 100)]public float hp = 100;
     [Range(1f, 20f)]public float speed = 3;
     public float atk = 1;
     public int gold = 0;
